+++ src/backend/joanie/core/api/admin/__init__.py
@@ -932,7 +932,9 @@ class BatchOrderViewSet(
         Once confirmed, the batch order's state transition to `to_sign`.
         """
         batch_order = self.get_object()
-        purchase_order_reference = request.data.get("purchase_order_reference")
+        purchase_order_reference = request.data.get(
+            "purchase_order_reference", ""
+        ).strip()
 
         if not batch_order.uses_purchase_order:
             raise ValidationError(
@@ -951,6 +953,9 @@ class BatchOrderViewSet(
             raise ValidationError(
                 _("Batch order's quote purchase order already confirmed.")
             )
+
+        if not purchase_order_reference:
+            raise ValidationError(_("Purchase order reference is required."))
 
         batch_order.quote.tag_has_purchase_order(
             purchase_order_reference=purchase_order_reference